[SCRUM - 57] - [WEB] Backend edycja statusu zgłoszenia

+++ pjaidApp/backend/pjaidJavaApp/src/main/java/org/api/pjaidapp/repository/TicketRepository.java
@@ -1,14 +1,14 @@
 package org.api.pjaidapp.repository;
 
-
 import org.api.pjaidapp.model.Ticket;
 import org.api.pjaidapp.model.TicketStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TicketRepository extends JpaRepository<Ticket, Integer> {
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
     List<Ticket> findByStatus(TicketStatus status);
-
     List<Ticket> findByTitleContainingIgnoreCase(String title);
+
+
 }
